+++ Gobchat.App/WebUIResources.Designer.cs
@@ -78,6 +78,24 @@ namespace Gobchat {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Chat background color.
+        /// </summary>
+        internal static string config_app_chatbox_backgroundcolor {
+            get {
+                return ResourceManager.GetString("config.app.chatbox.backgroundcolor", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Changes the background color of the chat. Don&apos;t go full transparent..
+        /// </summary>
+        internal static string config_app_chatbox_backgroundcolor_tooltip {
+            get {
+                return ResourceManager.GetString("config.app.chatbox.backgroundcolor.tooltip", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Collect information about character locations.
         /// </summary>
         internal static string config_app_ckb_actor {
@@ -213,24 +231,6 @@ namespace Gobchat {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show timestamp.
-        /// </summary>
-        internal static string config_app_ckb_showtime {
-            get {
-                return ResourceManager.GetString("config.app.ckb.showtime", resourceCulture);
-            }
-        }
-        
-        /// <summary>
-        ///   Looks up a localized string similar to .
-        /// </summary>
-        internal static string config_app_ckb_showtime_tooltip {
-            get {
-                return ResourceManager.GetString("config.app.ckb.showtime.tooltip", resourceCulture);
-            }
-        }
-        
-        /// <summary>
         ///   Looks up a localized string similar to Check for updates on start.
         /// </summary>
         internal static string config_app_ckb_update {
@@ -312,6 +312,15 @@ namespace Gobchat {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Larger.
+        /// </summary>
+        internal static string config_app_fontsize_larger {
+            get {
+                return ResourceManager.GetString("config.app.fontsize.larger", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Medium.
         /// </summary>
         internal static string config_app_fontsize_medium {
@@ -330,11 +339,29 @@ namespace Gobchat {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Smaller.
+        /// </summary>
+        internal static string config_app_fontsize_smaller {
+            get {
+                return ResourceManager.GetString("config.app.fontsize.smaller", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to .
         /// </summary>
         internal static string config_app_fontsize_tooltip {
             get {
                 return ResourceManager.GetString("config.app.fontsize.tooltip", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Very large.
+        /// </summary>
+        internal static string config_app_fontsize_verylarge {
+            get {
+                return ResourceManager.GetString("config.app.fontsize.verylarge", resourceCulture);
             }
         }
         
@@ -933,6 +960,15 @@ namespace Gobchat {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Chat tabs.
+        /// </summary>
+        internal static string config_main_nav_tabs {
+            get {
+                return ResourceManager.GetString("config.main.nav.tabs", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Copy this page from another profile.
         /// </summary>
         internal static string config_main_profile_copypage {
@@ -1360,6 +1396,141 @@ namespace Gobchat {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Add new tab.
+        /// </summary>
+        internal static string config_tabs_btn_newtab {
+            get {
+                return ResourceManager.GetString("config.tabs.btn.newtab", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Selected tab.
+        /// </summary>
+        internal static string config_tabs_tab_label {
+            get {
+                return ResourceManager.GetString("config.tabs.tab.label", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Apply mentions.
+        /// </summary>
+        internal static string config_tabs_tab_mention {
+            get {
+                return ResourceManager.GetString("config.tabs.tab.mention", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Apply rangefilter.
+        /// </summary>
+        internal static string config_tabs_tab_rangefilter {
+            get {
+                return ResourceManager.GetString("config.tabs.tab.rangefilter", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Apply roleplay formatting.
+        /// </summary>
+        internal static string config_tabs_tab_roleplay {
+            get {
+                return ResourceManager.GetString("config.tabs.tab.roleplay", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Show timestamp.
+        /// </summary>
+        internal static string config_tabs_tab_timestamp {
+            get {
+                return ResourceManager.GetString("config.tabs.tab.timestamp", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to .
+        /// </summary>
+        internal static string config_tabs_tab_timestamp_tooltip {
+            get {
+                return ResourceManager.GetString("config.tabs.tab.timestamp.tooltip", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Settings.
+        /// </summary>
+        internal static string config_tabs_tabtable_action_config {
+            get {
+                return ResourceManager.GetString("config.tabs.tabtable.action.config", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Delete .
+        /// </summary>
+        internal static string config_tabs_tabtable_action_delete {
+            get {
+                return ResourceManager.GetString("config.tabs.tabtable.action.delete", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Delete entry?.
+        /// </summary>
+        internal static string config_tabs_tabtable_action_delete_confirm {
+            get {
+                return ResourceManager.GetString("config.tabs.tabtable.action.delete.confirm", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Move one down.
+        /// </summary>
+        internal static string config_tabs_tabtable_action_down {
+            get {
+                return ResourceManager.GetString("config.tabs.tabtable.action.down", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Move one up.
+        /// </summary>
+        internal static string config_tabs_tabtable_action_up {
+            get {
+                return ResourceManager.GetString("config.tabs.tabtable.action.up", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Actions.
+        /// </summary>
+        internal static string config_tabs_tabtable_actions {
+            get {
+                return ResourceManager.GetString("config.tabs.tabtable.actions", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Name.
+        /// </summary>
+        internal static string config_tabs_tabtable_name {
+            get {
+                return ResourceManager.GetString("config.tabs.tabtable.name", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Visible.
+        /// </summary>
+        internal static string config_tabs_tabtable_visible {
+            get {
+                return ResourceManager.GetString("config.tabs.tabtable.visible", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Alliance.
         /// </summary>
         internal static string main_chat_channel_alliance {
@@ -1405,6 +1576,24 @@ namespace Gobchat {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to CWL1.
+        /// </summary>
+        internal static string main_chat_channel_crossworldlinkshell_1_abbreviation {
+            get {
+                return ResourceManager.GetString("main.chat.channel.crossworldlinkshell-1.abbreviation", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to CWL1.
+        /// </summary>
+        internal static string main_chat_channel_crossworldlinkshell_1_short {
+            get {
+                return ResourceManager.GetString("main.chat.channel.crossworldlinkshell-1.short", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to /cwl1.
         /// </summary>
         internal static string main_chat_channel_crossworldlinkshell_1_tooltip {
@@ -1419,6 +1608,24 @@ namespace Gobchat {
         internal static string main_chat_channel_crossworldlinkshell_2 {
             get {
                 return ResourceManager.GetString("main.chat.channel.crossworldlinkshell-2", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to CWL2.
+        /// </summary>
+        internal static string main_chat_channel_crossworldlinkshell_2_abbreviation {
+            get {
+                return ResourceManager.GetString("main.chat.channel.crossworldlinkshell-2.abbreviation", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to CWL2.
+        /// </summary>
+        internal static string main_chat_channel_crossworldlinkshell_2_short {
+            get {
+                return ResourceManager.GetString("main.chat.channel.crossworldlinkshell-2.short", resourceCulture);
             }
         }
         
@@ -1441,6 +1648,15 @@ namespace Gobchat {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to CWL3.
+        /// </summary>
+        internal static string main_chat_channel_crossworldlinkshell_3_abbreviation {
+            get {
+                return ResourceManager.GetString("main.chat.channel.crossworldlinkshell-3.abbreviation", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to /cwl3.
         /// </summary>
         internal static string main_chat_channel_crossworldlinkshell_3_tooltip {
@@ -1455,6 +1671,15 @@ namespace Gobchat {
         internal static string main_chat_channel_crossworldlinkshell_4 {
             get {
                 return ResourceManager.GetString("main.chat.channel.crossworldlinkshell-4", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to CWL4.
+        /// </summary>
+        internal static string main_chat_channel_crossworldlinkshell_4_abbreviation {
+            get {
+                return ResourceManager.GetString("main.chat.channel.crossworldlinkshell-4.abbreviation", resourceCulture);
             }
         }
         
@@ -1477,6 +1702,15 @@ namespace Gobchat {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to CWL5.
+        /// </summary>
+        internal static string main_chat_channel_crossworldlinkshell_5_abbreviation {
+            get {
+                return ResourceManager.GetString("main.chat.channel.crossworldlinkshell-5.abbreviation", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to /cwl5.
         /// </summary>
         internal static string main_chat_channel_crossworldlinkshell_5_tooltip {
@@ -1491,6 +1725,15 @@ namespace Gobchat {
         internal static string main_chat_channel_crossworldlinkshell_6 {
             get {
                 return ResourceManager.GetString("main.chat.channel.crossworldlinkshell-6", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to CWL6.
+        /// </summary>
+        internal static string main_chat_channel_crossworldlinkshell_6_abbreviation {
+            get {
+                return ResourceManager.GetString("main.chat.channel.crossworldlinkshell-6.abbreviation", resourceCulture);
             }
         }
         
@@ -1513,6 +1756,15 @@ namespace Gobchat {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to CWL7.
+        /// </summary>
+        internal static string main_chat_channel_crossworldlinkshell_7_abbreviation {
+            get {
+                return ResourceManager.GetString("main.chat.channel.crossworldlinkshell-7.abbreviation", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to /cwl7.
         /// </summary>
         internal static string main_chat_channel_crossworldlinkshell_7_tooltip {
@@ -1527,6 +1779,15 @@ namespace Gobchat {
         internal static string main_chat_channel_crossworldlinkshell_8 {
             get {
                 return ResourceManager.GetString("main.chat.channel.crossworldlinkshell-8", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to CWL8.
+        /// </summary>
+        internal static string main_chat_channel_crossworldlinkshell_8_abbreviation {
+            get {
+                return ResourceManager.GetString("main.chat.channel.crossworldlinkshell-8.abbreviation", resourceCulture);
             }
         }
         
@@ -1657,6 +1918,15 @@ namespace Gobchat {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to FC.
+        /// </summary>
+        internal static string main_chat_channel_guild_abbreviation {
+            get {
+                return ResourceManager.GetString("main.chat.channel.guild.abbreviation", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to /fc.
         /// </summary>
         internal static string main_chat_channel_guild_tooltip {
@@ -1671,6 +1941,15 @@ namespace Gobchat {
         internal static string main_chat_channel_linkshell_1 {
             get {
                 return ResourceManager.GetString("main.chat.channel.linkshell-1", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to L1.
+        /// </summary>
+        internal static string main_chat_channel_linkshell_1_abbreviation {
+            get {
+                return ResourceManager.GetString("main.chat.channel.linkshell-1.abbreviation", resourceCulture);
             }
         }
         
@@ -1693,6 +1972,15 @@ namespace Gobchat {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to L2.
+        /// </summary>
+        internal static string main_chat_channel_linkshell_2_abbreviation {
+            get {
+                return ResourceManager.GetString("main.chat.channel.linkshell-2.abbreviation", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to /l2.
         /// </summary>
         internal static string main_chat_channel_linkshell_2_tooltip {
@@ -1707,6 +1995,15 @@ namespace Gobchat {
         internal static string main_chat_channel_linkshell_3 {
             get {
                 return ResourceManager.GetString("main.chat.channel.linkshell-3", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to L3.
+        /// </summary>
+        internal static string main_chat_channel_linkshell_3_abbreviation {
+            get {
+                return ResourceManager.GetString("main.chat.channel.linkshell-3.abbreviation", resourceCulture);
             }
         }
         
@@ -1729,6 +2026,15 @@ namespace Gobchat {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to L4.
+        /// </summary>
+        internal static string main_chat_channel_linkshell_4_abbreviation {
+            get {
+                return ResourceManager.GetString("main.chat.channel.linkshell-4.abbreviation", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to /l4.
         /// </summary>
         internal static string main_chat_channel_linkshell_4_tooltip {
@@ -1743,6 +2049,15 @@ namespace Gobchat {
         internal static string main_chat_channel_linkshell_5 {
             get {
                 return ResourceManager.GetString("main.chat.channel.linkshell-5", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to L5.
+        /// </summary>
+        internal static string main_chat_channel_linkshell_5_abbreviation {
+            get {
+                return ResourceManager.GetString("main.chat.channel.linkshell-5.abbreviation", resourceCulture);
             }
         }
         
@@ -1765,6 +2080,15 @@ namespace Gobchat {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to L6.
+        /// </summary>
+        internal static string main_chat_channel_linkshell_6_abbreviation {
+            get {
+                return ResourceManager.GetString("main.chat.channel.linkshell-6.abbreviation", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to /l6.
         /// </summary>
         internal static string main_chat_channel_linkshell_6_tooltip {
@@ -1783,6 +2107,15 @@ namespace Gobchat {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to L7.
+        /// </summary>
+        internal static string main_chat_channel_linkshell_7_abbreviation {
+            get {
+                return ResourceManager.GetString("main.chat.channel.linkshell-7.abbreviation", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to /l7.
         /// </summary>
         internal static string main_chat_channel_linkshell_7_tooltip {
@@ -1797,6 +2130,15 @@ namespace Gobchat {
         internal static string main_chat_channel_linkshell_8 {
             get {
                 return ResourceManager.GetString("main.chat.channel.linkshell-8", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to L8.
+        /// </summary>
+        internal static string main_chat_channel_linkshell_8_abbreviation {
+            get {
+                return ResourceManager.GetString("main.chat.channel.linkshell-8.abbreviation", resourceCulture);
             }
         }