+++ docs/CHANGELOG.md
@@ -9,6 +9,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com)
 - Chat
   - It's now possible to change the background color of the chat in `Config / App`
 
+- Tabs
+  - They can be configurated in `Config / Chat tabs`
+
+- More font sizes
+  - Smaller, larger and very large
 
 ### Changed
 - Swapped some old JS Dialogs with custom made ones